+++ szconfig/doc.go
@@ -3,4 +3,4 @@ The g2config package is used make G2Config requests to a Senzing gRPC server.
 The gRPC definitions are at https://github.com/senzing-garage/g2-sdk-proto.
 The Senzing gRPC server is at https://github.com/senzing-garage/servegrpc.
 */
-package g2config
+package szconfig
