+++ szconfigmanager/g2configmgr_examples_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package g2configmgr
+package szconfigmanager
 
 import (
 	"context"
