+++ szconfigmanager/main.go
@@ -1,4 +1,4 @@
-package g2configmgr
+package szconfigmanager
 
 // ----------------------------------------------------------------------------
 // Constants
