+++ szdiagnostic/doc.go
@@ -3,4 +3,4 @@ The g2diagnostic package is used make G2Diagnostic requests to a Senzing gRPC se
 The gRPC definitions are at https://github.com/senzing-garage/g2-sdk-proto.
 The Senzing gRPC server is at https://github.com/senzing-garage/servegrpc.
 */
-package g2diagnostic
+package szdiagnostic
