+++ szdiagnostic/g2diagnostic_examples_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package g2diagnostic
+package szdiagnostic
 
 import (
 	"context"
