+++ szengine/doc.go
@@ -1,6 +1,6 @@
 /*
-The g2configmgr package is used make G2Configmgr requests to a Senzing gRPC server.
+The g2engine package is used make G2Engine requests to a Senzing gRPC server.
 The gRPC definitions are at https://github.com/senzing-garage/g2-sdk-proto.
 The Senzing gRPC server is at https://github.com/senzing-garage/servegrpc.
 */
-package g2configmgr
+package szengine
