+++ szengine/g2engine_examples_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package g2engine
+package szengine
 
 import (
 	"context"
