+++ szproduct/g2product_test.go
@@ -1,4 +1,4 @@
-package g2product
+package szproduct
 
 import (
 	"context"
