+++ szproduct/main.go
@@ -1,4 +1,4 @@
-package g2product
+package szproduct
 
 // ----------------------------------------------------------------------------
 // Constants
